[Enhance] Separate installation for each tasks on release task (#1869)
* separte_import

* align with pre commit

* update unit test code

* add separate task env pre  merge test & aplly it to github action

* add multiprocess to requirement
