+++ tox.ini
@@ -50,6 +50,95 @@ commands =
     coverage report -m --fail-under=0
     coverage xml -o {toxworkdir}/coverage.xml
 
+[testenv:pre-merge-cls]
+deps =
+    torch @ https://download.pytorch.org/whl/cu117/torch-1.13.1%2Bcu117-cp38-cp38-linux_x86_64.whl
+    torchvision @ https://download.pytorch.org/whl/cu117/torchvision-0.14.1%2Bcu117-cp38-cp38-linux_x86_64.whl
+    torchtext @ https://download.pytorch.org/whl/torchtext-0.14.1-cp38-cp38-linux_x86_64.whl
+    mmcv-full @ https://download.openmmlab.com/mmcv/dist/cu117/torch1.13.0/mmcv_full-1.7.0-cp38-cp38-manylinux1_x86_64.whl
+    -r{toxinidir}/requirements/api.txt
+    -r{toxinidir}/requirements/base.txt
+    -r{toxinidir}/requirements/openvino.txt
+    -r{toxinidir}/requirements/classification.txt
+    -r{toxinidir}/requirements/dev.txt
+use_develop = true
+commands =
+    coverage erase
+    coverage run -m pytest -ra --showlocals --junitxml={toxworkdir}/test-results.xml tests/integration/cli/classification
+    coverage report -m --fail-under=0
+    coverage xml -o {toxworkdir}/coverage.xml
+
+[testenv:pre-merge-det]
+deps =
+    torch @ https://download.pytorch.org/whl/cu117/torch-1.13.1%2Bcu117-cp38-cp38-linux_x86_64.whl
+    torchvision @ https://download.pytorch.org/whl/cu117/torchvision-0.14.1%2Bcu117-cp38-cp38-linux_x86_64.whl
+    torchtext @ https://download.pytorch.org/whl/torchtext-0.14.1-cp38-cp38-linux_x86_64.whl
+    mmcv-full @ https://download.openmmlab.com/mmcv/dist/cu117/torch1.13.0/mmcv_full-1.7.0-cp38-cp38-manylinux1_x86_64.whl
+    -r{toxinidir}/requirements/api.txt
+    -r{toxinidir}/requirements/base.txt
+    -r{toxinidir}/requirements/openvino.txt
+    -r{toxinidir}/requirements/detection.txt
+    -r{toxinidir}/requirements/dev.txt
+use_develop = true
+commands =
+    coverage erase
+    coverage run -m pytest -ra --showlocals --junitxml={toxworkdir}/test-results.xml tests/integration/cli/detection
+    coverage report -m --fail-under=0
+    coverage xml -o {toxworkdir}/coverage.xml
+
+[testenv:pre-merge-seg]
+deps =
+    torch @ https://download.pytorch.org/whl/cu117/torch-1.13.1%2Bcu117-cp38-cp38-linux_x86_64.whl
+    torchvision @ https://download.pytorch.org/whl/cu117/torchvision-0.14.1%2Bcu117-cp38-cp38-linux_x86_64.whl
+    torchtext @ https://download.pytorch.org/whl/torchtext-0.14.1-cp38-cp38-linux_x86_64.whl
+    mmcv-full @ https://download.openmmlab.com/mmcv/dist/cu117/torch1.13.0/mmcv_full-1.7.0-cp38-cp38-manylinux1_x86_64.whl
+    -r{toxinidir}/requirements/api.txt
+    -r{toxinidir}/requirements/base.txt
+    -r{toxinidir}/requirements/openvino.txt
+    -r{toxinidir}/requirements/segmentation.txt
+    -r{toxinidir}/requirements/dev.txt
+use_develop = true
+commands =
+    coverage erase
+    coverage run -m pytest -ra --showlocals --junitxml={toxworkdir}/test-results.xml tests/integration/cli/segmentation
+    coverage report -m --fail-under=0
+    coverage xml -o {toxworkdir}/coverage.xml
+
+[testenv:pre-merge-action]
+deps =
+    torch @ https://download.pytorch.org/whl/cu117/torch-1.13.1%2Bcu117-cp38-cp38-linux_x86_64.whl
+    torchvision @ https://download.pytorch.org/whl/cu117/torchvision-0.14.1%2Bcu117-cp38-cp38-linux_x86_64.whl
+    torchtext @ https://download.pytorch.org/whl/torchtext-0.14.1-cp38-cp38-linux_x86_64.whl
+    mmcv-full @ https://download.openmmlab.com/mmcv/dist/cu117/torch1.13.0/mmcv_full-1.7.0-cp38-cp38-manylinux1_x86_64.whl
+    -r{toxinidir}/requirements/api.txt
+    -r{toxinidir}/requirements/base.txt
+    -r{toxinidir}/requirements/openvino.txt
+    -r{toxinidir}/requirements/action.txt
+    -r{toxinidir}/requirements/dev.txt
+use_develop = true
+commands =
+    coverage erase
+    coverage run -m pytest -ra --showlocals --junitxml={toxworkdir}/test-results.xml tests/integration/cli/action
+    coverage report -m --fail-under=0
+    coverage xml -o {toxworkdir}/coverage.xml
+
+[testenv:pre-merge-anomaly]
+deps =
+    torch @ https://download.pytorch.org/whl/cu117/torch-1.13.1%2Bcu117-cp38-cp38-linux_x86_64.whl
+    torchvision @ https://download.pytorch.org/whl/cu117/torchvision-0.14.1%2Bcu117-cp38-cp38-linux_x86_64.whl
+    torchtext @ https://download.pytorch.org/whl/torchtext-0.14.1-cp38-cp38-linux_x86_64.whl
+    mmcv-full @ https://download.openmmlab.com/mmcv/dist/cu117/torch1.13.0/mmcv_full-1.7.0-cp38-cp38-manylinux1_x86_64.whl
+    -r{toxinidir}/requirements/api.txt
+    -r{toxinidir}/requirements/base.txt
+    -r{toxinidir}/requirements/openvino.txt
+    -r{toxinidir}/requirements/anomaly.txt
+    -r{toxinidir}/requirements/dev.txt
+use_develop = true
+commands =
+    coverage erase
+    coverage run -m pytest -ra --showlocals --junitxml={toxworkdir}/test-results.xml tests/integration/cli/anomaly
+    coverage report -m --fail-under=0
+    coverage xml -o {toxworkdir}/coverage.xml
 
 [testenv:build-doc]
 deps =
